use vsp client from internal pkg for purchasing tickets
Maintain a record of vsp clients connected to a wallet to track
and process all tickets purchased with a VSP using the wallet.
This record is maintained in the loader package upstream but that
doesn't fit well into the multiwallet architecture since each vsp
client is tied to a single wallet.