spv syncer: maintain different sidechains for each wallet
The sidechain acts as a temporary storage location for blocks that
are fetched from a remote peer, prior to being added to each wallet's
mainchain. While unlikely, but especially in the case of a reorg, the
data in the sidechain might become inconsistent because it depends on
wallet-specific information such as blocks already existing in each
wallet's mainchain and blocks that become re-orged from a wallet.

This change ensures that the sidechain data remains consistent for each
wallet.